"5:20pm. Ok, I have the line and the parabola. Let me try solving this using linear equations.
5:35pm.

```
In[173]:= coef[x_, len_] := Table[If[i == 0, 1, x^i], {i, 0, len - 1}]

In[172]:= linearEqsLineData =
 Table[coef[data[[i]][[1]], 2], {i, Length[data]}]

Out[172]= {{1, 0}, {1, 1}, {1, 3}, {1, 5}}

In[174]:= linearEqsLineTarget =
 Table[data[[i]][[2]], {i, Length[data]}]

Out[174]= {1, 0, 2, 4}
```

Somehow this is taking me a while. This code is ugly, no doubt about it. But that will always be that case for non-ML languages.

5:40pm. Focus, let me see what the linear solver does with this.

5:50pm.

```
In[187]:= LeastSquares[linearEqsLineData, {t1, t2, t3, t4}]

Out[187]= {(35 t1)/59 + (26 t2)/59 + (8 t3)/59 - (10 t4)/
  59, -((9 t1)/59) - (5 t2)/59 + (3 t3)/59 + (11 t4)/59}
```

I figured out how to do this. `Solve` and `LinearSolve` do nothing, but least squares is the one that I need here.

Now that I have this, I can get to work in putting in the stuff from the book.

Let me do that and then I am done for the day.

6:05pm. I am just looking at the problem and thinking. It can't possibly be true that the thing is correct.
`sum j (inl j -> A i j) = T i`?

I really have an issue with that `i` there. Nowhere is it given, but it is possible to infer what the matrix of it should be.

Well, let me construct the dataset for this.

6:50pm. I am out of time here. I'll do this thing tomorrow.

I'll make better progress from here on out. I won't bother with making precise formal proofs of these exercises right now. I'll leave that for Coq on the second iteration.

Now, let me have some fun."