Let me start the day. Today I will finally deal with the first exercise. It will be a long way from here to dealing with the credit assignment problem, but the steps need to be made.
Now that I am doing more concrete work my state of mind is better.

Well, let me play around for a bit. That is the only way I'll be able to get through this.

Though there is zero chance of the equations in the first exercise being false, I am really skeptical that the thing will work. I mean, I sure wouldn't come up with this on my own.

Regarding the size of the matrix, I remember from the linear algebra lectures of years ago that the number of equations needs to be at least the number of variables.

So for 2 variables which I have now, I should make the number of rows of the A matrix 2.

With that out of the way, let me make that dataset.