Let me stop to do the chores.
```
// The third exercise from the Pattern Matching and Machine Learning book.

type Box =
    | Red
    | Green
    | Blue

type Fruit =
    | Orange
    | Lime
    | Apple

let normalize_ints x =
    let total = Array.fold (fun s (_,num_items) -> s + num_items) 0 x
    Array.map (fun (dsc,num_items) -> dsc, float num_items / float total) x

let boxes =
    [|
    Red, normalize_ints [|Apple, 3; Lime, 3; Orange, 4|]
    Blue, normalize_ints [|Apple, 1; Lime, 0; Orange, 1|]
    Green, normalize_ints [|Apple, 3; Lime, 4; Orange, 3|]
    |]
```

I had to think about this for a while. I actually got confused at this point as to how to represent the distribution of distributions. I know how now. I'll do it when I get back and implement inference via enumeration.

This is how Bayesian inference should be done - using probabilistic programming. I was going to do it directly by hand, but for the first time I am realizing - it is not that math is hard. It is that it literally makes zero sense. It is just a bunch of equations floating about that only make sense in the author's head. In contrast, explaining something to a machine is a much richer specification.

This is how I should be done it - I need to elevate math to the level of programming, not programming to the level of math.

I am going to go through this book and then master formal methods after that. At the very least, that will give me confidence that I am not lacking as a practitioner.