I got up 2h ago, but had to do some replying. I am almost ready to start now. Actually, I think I will just have breakfast a bit earlier and then do that rather than risk having my focus broken just as I attain it.
I had time to think about the exercise I did yesterday over the night. I really do get it now.

The last 10% I do not get will be dealt with once I visualize the examples in the book on my own.

That is what will be my focus for the first part of the day. I'll move on to the following exercises after that.

Right now I am just about done with 1.4, but the exercises go up to 1.40. It will take me quite a while to go through this and after that 13 more chapters await me no doubt filled to the brim with them as well.

The only hope is that I bridge the gap and start internalizing math in the same way I do programming.

If I can do that it will give me a fighting chance in this quest for the Singularity.

It is not enough to wait for the algorithm to come to me - I need to go out there and get it myself. I cannot allow the others to have advantages in the conceptual realm. The holes need to be filled and the truth of hierarchy must be unveiled.