Let me resume.
Before I do that, let me say a few words extra. It is good that I was thorough here as I would have no idea that there is an inverse relationship between moving a variate and moving a probability distribution. Meaning there is a connection between mapping a sample from a distribution and mapping the distribution itself. I'd never have guessed that a functional inverse would play such a prominent role here.

If I just went through that part speedily, I'd completely missed the whole point of the segment.

Really, damn the book. What is even the point of writing that out if you are not going to highlight the use for the change of variable. At best that part could be considered half done.

I am going to have to be on my guard.

With that out of the way, let me return to the book. Before I got distracted by this yesterday, last time I was getting ready to chase down those old references. Let me do that now.