"5:10pm. Hmmmm...let me go over the stuff I do not understand.
1) Why is he suddenly talking about the Jacobian of the change of variables? Sure, it is a nice observation, but is there a point is saying it here?

2) I have absolutely no idea what is going on in (9) and (10). Also, it does seem like he really did mean I^2. how confusing.

3) I have absolutely no idea how `y = x - y` propagates.

I might be able to figure out the last two points if I could figure out what change of variable is actually doing here.

5:15pm. I'll put the book on pause and go through the integration section on Khan Academy. I am decently familiar with differentiation at this point thanks to my deep learning work, but it is obviously I need to refresh my memory of its various aspects.

I should be able to figure out what it going on in this exercise after I am done with my review.

Pretty much everything going forward is going to be about integration so there is no choice, but to get good.

Let me do the chores now and then I'll put the last bit of work in for the day."