"12:10pm. Ah, it has been a while since I did the quizzes last. The integration section had definitely been expanded since then.
12:35pm. Let me do the quiz.

12:45pm. Damn, I keep leveling down on the quizzes.

12:55pm. Focus me, let me do the first one properly.

I really dig the feedback that Khan Academy gives me by the way. It is kind of a pain to have to do these calculation in my head, and it makes it a lot easier.

1pm. Er, what the hell. I selected the correct one and got a fail.

Agh, the question was asking me to select more than one. Whops.

1:30pm. Ugh, I can't believe that this took me a whole hour. Let me move on to the next part.

1:35m. Focus me. Move on to the next thing. 4.5 more hours of this. Doing these quizzes is kind of like a game so I can do it.

2:25pm. I am not going to be able to do this in a single day. No way. It will take me two days to get through the lectures and the quizzes and probably more than that once I add the unit test to the mix.

2:45pm. ...I am on the wrong track here. These tests and quizzes are time consuming, but they are not increasing my understanding of the subject.

3h of this has been enough. Let me just skim the lectures until I get to the areas of interest. I do not have the time to mess around with this.

2:50pm. https://youtu.be/pWtt0AvU0KA

Proofs like these are what I need. This is what I need to make my own.

2:55pm. https://youtu.be/pWtt0AvU0KA?t=306

Ah, this is brilliant. I was really wondering why the fundamental theorem of calculus works. I thought it was strange that it ignored one of the bounds, but here it all comes together.

3:25pm. Had to take a little break. Yeah, I got distracted by those cutesy problems that I really did not need to deal with them.

What I need now is an intuitive understanding of the proof techniques, like in the last video.

Actually, I can't believe I could not figure this out on my own until it was shown to me. I really do have a lot to learn before I can even consider myself a mediocre mathematician.

Now let me watch the other proof.

What I really need to cover though is trig integration before the day is over. I should be able to make this in the next 2.5 hours."