I got up late again and decided to have an early breakfast. I now have 7 clear hours ahead of me in which to study.
Let me do it. I did some thinking over the night, but I am still uninspired as to how integrate that exponential function.

For the next two hours, let me just go through all the videos and then I will take a break and reevaluate where I stand.

In that case that I still do not know how to do it after that, no doubt this kind of problem has a detailed explanation on the net somewhere.

As expected, I am really hilariously bad at math. But it won't be that way forever. Whether it be doing proofs in DT languages or algebra,  I will improve my skills.