"11:15am. a^2-x^2=x=a*sin theta
Well in the exponent, I could leave a^2 as 0.

...No, that would make everything undefined. I do not really get it.

11:20am. Actually, I might be able to substitute the `o` for something I can simplify and then substitute it back...

Yeah, that avenue might be worth looking into. Let me continue with the lectures. As expected, my focus is quite high in the morning.

Edit: No, this is dumb. I do not want to constrain `o` to anything.

12:05pm. I am not sure why he did not just substitute theta directly in that last problem, but nevermind. Let me take a break here and then I will go through the last section.

12:30pm. Let me resume. Time to go through the last section. I am realizing that I won't be able to quite internalize this until I do some exercises with this, so maybe I will do the quizzes after I am done with the last section.

1:25pm. I am done with all the videos. That having said, I am still not sure how I would integrate `exp (neg x^2)`. My skills really are not enough yet.

Though I said I wanted to skip the quizzes and the exercises, how about I do them after all?

It is not that there is that much basic math to go around so I might as well dedicate some time to mastering integration properly.

1:30pm. Agh, if it is not one thing then it is another. Let me do it. I'll do all the exercises, the quizzes and then do the unit test and get my complete masteries.

That will give me assurance that I've covered the basics and can go look at the solutions for the problem once again."