"5:40pm. Done with lunch. Let me watch the video on proof 5. I have a little bit left before I call it a day.
This whole thing of trying to draw power from math is bound to turn out bloody for me, but I have to go through the motions.

https://youtu.be/tGn9EBnMKJc
Gaussian Integral 5 Volumes

5:45pm. Ok, I get why V goes from 0 to 1 in the integral. `z` really is 1 when `x^2` and `y^2` are 0. When `x` and `y` are zero, the z is at it maximum.

Agh, I am such a fool. I am not using my brain at all. What confused me here is that I was thinking about the normalized Gaussian, rather than the `exp (-x^2)` one. I remained stuck in my mind that the regular Gaussian has a maximum at 0.4. What a blunder.

Well, now that I've debugged this mental error I can move on.

Thankfully this geometric proof does not have a lot of symbolic manipulation behind it so I'll be able to make headway.

5:55pm. https://youtu.be/tGn9EBnMKJc?t=379

This geometric proof is great. It is probably the simplest possible variant there is.

But is it really necessary to do the integration in the third dimension. Would it not be possible to switch axes?

```
y = exp (-x*x)
log y = -x*x
-log y = x*x
- sqrt (log y) = x
```

Now what would be the integral of `log y`?

...Well, that Erf comes in again.

But actually if I do not take the sqrt...then the pi is missing...

6:15pm. Enough of this. I do not have the mental energy to try and go the sixth proof and the rest today. I'll leave them for tomorrow.

I still cannot really understand why it is necessary to introduce extra dimensions and change the problem formulation in order to get the right insight. I'll put this in the back of my bind as I go along. Maybe at some point I will realize the answer.

6:20pm. But for now, another day is over and it is time for me to get some rest."