"11am. I see I missed a minus there...
Mathematica could be a lot more useful as a proof assitant than it is now.

```
I^2 =
  In (from: 0 to: inf) <|
    In (from: 0 to: inf) <|
      exp (-(x^2 + y^2)) * dx * dy
```

Now let me make a the substitutions.

```
x = t * y
dx = y * dt

I^2 =
  -0.5 *
    In (from: 0 to: inf) <|
      In (from: 0 to: inf) <|
        exp -(y^2 * (t^2 + 1)) * y * dy * dt
```

Now the paper claims that `In (from: 0 to: inf) <| y * exp -(a*y^2) * dy = 1 / 2a` for `a > 0`. Is that really true? Let me make sure of it.

```
In (y * exp -(a*y^2) * dy) = -0.5 * exp (-a * y^2) / a
```

This is the indefinite integral. I can see plugging `inf` into `y` would result in it being 0. And result in it being `1 / 2 a` when plugging in 0.

Still what about plugging in `-inf`? If `y = -inf` then what should `y^2` be? Well according to Mathematica it should be `inf`. Then of course the definite integral would be 0 and the whole expression would be useless.

11:20am. I am just in a daze here. The way those infinities behave is just so arbitrary. On the surface, the thing is about algebraic manipulation, but underneath it seems to be about constraint satisfaction.

Is this really math at its core?

Yeah, it is not wonder I am bad at this.

11:30am. It is disappointing that I could not replicate the substitution proof even after seeing it before, but let me just go on.

Let me take a look at proof six and then I will have breakfast.

12:20pm. Let me resume. I want to go through the video for proof 6. I am not familiar with the gamma function, so no doubt it will have extra insight."