"Let me take a look at proof two again.
```
In (from: 0 to: inf) (y * exp (-a*y^2)) dy
```

Why is this 1/2a?

Let me substitute u for -a*y^2.

```
u = -a*y^2
du = -a*2*y*dy

bounds:
  (from: (-a*0) to: -a*inf^2)
  (from: 0 to: -a*inf)
  (from: 0 to: -inf) // For a > 0

In ((from: 0 to: -inf) (exp u / (-a*2)) du
exp u / (-a*2) | (from: 0 to: -inf)
(exp -inf / (-a*2)) - (exp 0 / (-a*2))
0 - (1 / (-a*2))
-1 / (-a*2)
1 / (a*2)
```

9:20am. I actually made a mistake in assuming that the bounds are unchanged while writing the above...

I've done some reflecting last night and I've come to a conclusion regarding my math attempts.

The reason why I constantly make mistakes is because I have a serious lack of discipline. It seems 12 years of schooling could not hammer home the lesson to not take mental shortcuts. Pretty much 100% of the mistakes in the past two weeks is because I was lazy and did not feel writing every step out explicitly.

Yeah, it is true - having a programming language that could verify all my steps would be a boon. But I should not need it for short math fragments such as these.

A proof assistant for these kinds of simple proofs would be overkill. I just need to increase my focus instead and stop being such a donkey."