"8:30am. "Could you try with the latest CUDA 10.1 and see if you will still hit the issue?"
This is for that tuple compilation bug from a few years ago. I am going to skip doing this as I do not feel like messing with it. It would take me hours of backtracking. Presumably they can fix this themselves. I'll get to it when I feel like it which might be years down the road.

8:40am. Let me do my morning reading and then I will try going through the first proof again.

Somehow though I am frittering time by the truckload, I do not at all feel like hurrying my pace.

The way I am going now is good enough. That is the sense I am getting."