"3:20pm. I am back. I do not feel like resuming right away.
3:55pm. Let me resume. Time to finish the paper.

4:15pm. May will come in a week. I should start thinking of what I am going to write in the monthly review.

4:40pm. Oh, just how did I spend so much paper reading that expository paper.

I really want to finish the proof videos. Let me go through the Stirling one. Then the Fourier one after that. I'll leave Coq for later.

5:05pm. My focus is shot at this point and I couldn't really focus on the video. To be honest, it is just a long boring sequence of straightforward calculation anyway. I won't bother myself over trying to internalize the Stirling GI proof.

I guess I'll give the Fourier video a shot and then call it a day.

Today was mostly failure and embarrassment over not being able to do those calculations, but the Coq library that was introduced to me while not an immediate asset does make the prospect of learning Coq a lot more appealing now that I know it can grok advanced mathematics. I was pessimistic on its ability to be useful in this domain, but I am glad I was wrong.

Though I expect even if I get good at this that the formal proofs will be tedious, but this is the sort of thing that will get better at time goes on and the technology improves.

5:10pm. Now, let me try Fourier.

https://www.youtube.com/watch?v=Tztpvz0ukG8
Gaussian Integral 10 Fourier Way

I'll leave studying the contour integration for tomorrow after which I will finish the last two proofs. Thankfully they are not calculation heavy.

After that I still need to figure out how the change of variable is being done in the first proof.

Once that is done, I will be done with exercise 1.7 and will be able to move on to the next thing in the ML book.

5:15pm. No forget it. I am done. I am not paying attention at all to what he is saying at this point.

It will just have to be done tomorrow.

5:25pm. https://www.reddit.com/r/MachineLearning/comments/bgvefd/d_have_we_hit_the_limits_of_deep_reinforcement/
https://www.reddit.com/r/MachineLearning/comments/bfq8v9/d_openai_five_vs_humans_currently_at_410633_992/

Yeah, I'd rather just waste time at this point.

I am done for the day.

I essentially need to make a paradigm shift in RL to get what I want which is impossible. So at best I will be wondering around and learning random things. There is no intricacy to my plan going forward - I will just build my skills and that is it.

Without a doubt, thanks to all the math the next few years are going to be hell, but there might be some silver lining to be found in dependently typed programming. And thankfully math builds on itself. If I can get the computer to help me out a little, I might be able to push my capabilities beyond what I could get if I did it with just pen and paper. Math is ripe for an overhaul by this point.

And theorem proving seems to presage developments in programming languages, so learning Coq won't be a waste of time in that sense either. By the time dependent types hit the mainstream, they will be significantly more ergonomic that they are today. And maybe, just maybe I will find some ideas on how to improve Spiral.

I do hope that I will find some thing in the next 3 years that will allow me to enter the kind of profession that I desire. I definitely did not intend to become a researcher, but I will become one if that is what is needed."