"1:55pm. https://youtu.be/VVPu5fWssPg?t=3648
In this part I am reallying wondering whether it would be possible to set `sqrt (1 - x^2/4 - y^2/9) = 0` in the integral.

...Ah, I do not get it. Maybe I should watch this whole lecture more carefully rather than just skimming. I really am lacking so much in the basics after all.

2:10pm. https://youtu.be/VVPu5fWssPg?t=4432

This seems like it might be interesting. Hopefully this will finally tell me how to do the transformation of the regions.

2:35pm. https://www.youtube.com/channel/UCoHhuummRZaIVX7bD4t2czg

Well, no, but here is Professor Leonard's channel. Let me take a look at his playlist.

2:40pm. https://www.youtube.com/watch?v=9FtHB7V14Fo&list=PL5102DFDC6790F3D0

Ugh, I am going to need to go through this at some point. Not now though.

2:45pm. Shit his videos are long. I could be watching this stuff for the next month.

https://www.youtube.com/watch?v=sWUyFQQ5QeI
Calculus 2 Lecture 10.4: Using Polar Coordinates and Polar Equations

I think I should be focusing on the 14.x series, but let me skim this for 5m or so.

2:50pm. No that video is not what I am looking for. But damn the amount of math puzzles really is infinite is it not.

Thankfully that actual concepts that I need to master is much less.

I think once I get a proper grasp on the areas of science that I need, the way to figure out the missing piece that I need to do credit assignment properly might show itself.

Let me take a look at the 14.x series of videos.

2:55pm. https://youtu.be/HxRG_phgGUw?list=PLDesaqWTN6ESk16YRmzuJ8f6-rnuy0Ry7&t=67

This video is literally 3:23 hours long. Holy crap. And he starts it off by talking about R and how his calc teacher never explained it. I am getting near to what I need. The next video seems to be on the polar regions, so that is the one I should watch instead. I'll skim this one for a bit though.

3:05pm. https://www.youtube.com/watch?v=HA41kYxVYnw&list=PLDesaqWTN6ESk16YRmzuJ8f6-rnuy0Ry7&index=24
Calculus 3 Lecture 14.3: Double Integrals over POLAR REGIONS

Finally polar regions. Let me watch this for a bit.

3:25pm. https://youtu.be/HA41kYxVYnw?list=PLDesaqWTN6ESk16YRmzuJ8f6-rnuy0Ry7&t=6448

I had enough of this.

I guess transformations to polar coordinate are intended to be intuitive to some degree rather than straightforward algebraic. With the Gaussian integral I can easily see that the transformation given in the solutions is correct. And I get the Jacobian determinant for the change of variables as well.

Ahh, finally done with this. I think I will resume the ML book. Let me take a break first."