https://stackoverflow.com/questions/55933706/is-it-possible-to-prove-implies-to-or-de-morgan-not-and-not-without-resorti
I am crazy right now. I said that I never lost sleep because of math. Well, that sure happened now.

Actually I am kidding. I slept quite well, but I jumped straight at the problem at 8am as soon as my eyes opened. And I kept at it till 9pm yesterday.

I think I had a steady pace for the past week, but I got so hung up on this problem that I am going to have to tie myself to a post in order to keep myself from working on it.

Ah this thing...I had such a nice flow, but now I am berserk. I need to move on to the next chapter.

And make that post on the PL monthly thread.

I really need to pare it down to two posts somehow rather than writing novels every time.

Hopefully I'll get something from that SO question and that will give me closure. I won't waste my energy on it anymore.

I am still too green. I need to learn from this experience.

The only thing good about this is that this kind of frenzy is actually good for learning. It is certainly better to be like this than disinterestedly watching lectures. You can be sure that I am going to internalize the skills attained in this fashion. And that is one of the main points of this whole thing.