Let me stop slacking and start doing. I solved everything during the night.
Mentally, I seem to be doing dependently typed programming in Idris. But the amazing thing is how trivial the proofs for both functions are. Just how is it possible that I got stuck for hours yesterday when either of the proofs should have taken a minute at most.

Let me figure this out.

As expected, programming really starts when you determine what should be. I really gained a lot from the Little Typer and the TDD in Idris book.