`reg_exp_of_list_spec` was stupidly tedious and long winded. Let me call it a day here.
Maybe at this rate it will take me more than a month to get to the end of the book especially if I am aiming for completion. But I am getting better at this, so the pace should pick up. I've really spent a lot on just a few problems in the past few days. Once I get to the usual conveyor belt rate, things will get moving.

Let me check out how much is left...

I am a bit past third of the way in the first volume. Completing it should be doable in the next two or three weeks.

Either way, it is fine if it takes me an extra month or two to do this - I want the ability to do these kinds of proofs with proficiency.

Now that I've decided that it makes more sense to get a job than mess with tabular algorithms, that sort of gives me only two options.

1) Get a job.
2) Crack credit assignment.

Life is just too special to waste it on the first option. I'd rather aim big and be a failure than play it safe.

I do not know when it was that I decided that I would live this way, but for as long as I can remember that is what I strove for. Ultimately, there is only a finite amount of work. Eventually I will get through it and reap the benefits.

I pay lip service to sex and money, but I guess I do have something in me that will push me through years and years of self improvement. It will be good training for the actual loop.

In the loop, it is not like it itself will give me power. Rather it will just be an enabler of change. Just like I am doing now, the great power that I will attain will be taken by my own hands - even if it kills me. I do not need courage, nor do I have power at this juncture. I just need to keep the flame of my desire burning.

I do feel bad for putting Simulacrum on hiatus for so long, but writing would straight out force me into a job just to pay the editors and cover artist and all that shit. I am not sure there is even an audience for it.

And ultimately it is not that kind of story. The first three arcs, were a declaration of victory - a proof that I understand the essence of self improvement. The stuff after that and the stuff I am doing now is just me killing my weak self; without a good kick in the ass I would never have started moving. Though I have the outline for the next 3 arcs in my head, they need to be more than just a story. A suitable declaration of victory would be me proving that I understand the principles of intelligence along with it.

That would get me fired up.

Let me stop here for the day. Tomorrow, I will resume my zealous search for the mathematical handle.