"2:40pm. I had to think for a good long while exactly what is going on here.
The issue is simply that Coq sucks. It sucks a lot. The tactics suck. I would never run into the problems like I am doing now in Idris or Agda.

Here is the issue.

```
| MStar0 : forall re, exp_match [] (Star re)
| MStarApp : forall s1 s2 re,
               exp_match s1 re ->
               exp_match s2 (Star re) ->
               exp_match (s1 ++ s2) (Star re).
```

Remember these two constructors.

2:45pm.

```
Lemma MStar'' : forall T (s : list T) (re : reg_exp),
  s =~ Star re ->
  exists ss : list (list T),
    s = fold app ss []
    /\ forall s', In s' ss -> s' =~ re.
Proof.
intros.
dependent induction H.
- exists [[]].
  split.
  + reflexivity.
  + compute.
    intros.
    destruct H.
    * rewrite <- H.
```

```
1 subgoal
T : Type
re : reg_exp
s' : list T
H : [ ] = s'
______________________________________(1/1)
[ ] =~ re
```

What is missing here. The fact that the premise is...

```
| MStar0 : forall re, exp_match [] (Star re)
```

When I destructured `re` here and it asked me to prove the EmptySet it really confused me.

But if I had the premise it would be easy to show that `exp_match [] (Star EmptySet)` is an absurdity in the premise. So the only thing I could construct here would be the empty star and the empty string.

The thing is, the `remember` tactic which is the first thing I tried outright forces the induction to fail. Nonetheless, the `assert (Hbak := H)` should do the trick instead."