"2:55pm. No, `[] =~ Star EmptySet` is not an absurdity.
I am beyond confused here.

3pm. Ah, fuck.

I get it. I passed in an empty list of empty list an an example, when I really should have done [] instead.

The issue with it is that the outer list has in fact a single element due to that.

Uah...

Well, let me continue..."