I've finally gone through MStar''. As it turns out the reason why it took so long is because I missed a dumb rewrite and not the thing I thought up during the night. Ohhhhhhhhhhhh...
What could it be, whatever could it be...and it turned out to be that. The second branch turned out quite simple in th end.

The next thing on the list would be the 5 star problem. But since I never go to do it yesterday, let me take a break right now. It is time I had breakfast anyway.

Well, hopefully by the end of the month I will get better at this and work out all the kinks in my brain that cause me to do stupid things.