"https://mangadex.org/title/36767/the-villainess-will-crush-her-destruction-end-through-modern-firepower
Whether it be the smug grin, the silly small hat, or cute and stylish loli dual wielding weapons while doing the Duke Nukem pose the Astrid cover really caught my attention. It just oozes style. The ideal of inhumanity is the appealing outside...

```
“Then let me ask you this, Red Devil! Aren’t you merely a child? Killing countless people without a care in the world… Even women and children… These soldiers just want to return home to their families, how can you be so calm through all this!?”

The old man’s words were boring.

“It’s fine. There are certain synapses in the brain that cause one to be averse to murder. It’s these synapses that force this so-called ‘conscience’, the single biggest obstacle to killing. But, what if I could force them to stop working?”

Well, it’s actually something I’m pretty proud of.

“Such a thing, that’s impossible! You messed with your brain using magic? To disable your own conscience!?”

“That’s exactly right. I don’t feel pangs of conscience, mercy nor pity. I am merely a machine that mows down her enemies. If the opponent is an enemy soldier, I can pull the trigger without moral qualms. I don’t feel any compassion for the collateral damage either.”

I had finally achieved it. The human brain perfectly suited to combat.
```

...and the soul of a killing machine on the inside.

This just hits all the right notes for me. I am cautiously optimistic that this might turn out to be good.

5/6/2019

7:25am. I am up and fully rested. I actually figured out how to do the case last night. I got pretty much all the pieces yesterday, I just need to connect them today. It is truly amazing how nice it is when everything comes together. I feel like I understand what the lemma is right now too.

I am going to have to prove the auxiliary that I started yesterday in order to complete it, but now that I know its place, I'll be able to dedicate myself to it.

I'll go into it later. Let me chill a bit with Tog and Astrid. I never got the chance to read it yesterday as I was too tired after finishing work and went to bed."