"8:05am. https://stackoverflow.com/questions/56009771/how-to-split-the-length-inequality-hypothesis-in-the-pumping-lemma/56013413#56013413
Geh, I got a perfect answer. Even if I did not get this answer I just finished reasoning out how if I have a function of type...

```
  s =~ re ->
  pumping_constant re <= length s ->
  exists s1 s2 s3,
    s = s1 ++ s2 ++ s3 /\
    s2 <> [] /\
    forall m, s1 ++ napp m s2 ++ s3 =~ re.
```

Then it should be easy to prove...

```
  s =~ re ->
  pumping_constant re + a <= length s + b->
  exists s1 s2 s3,
    s = s1 ++ s2 ++ s3 /\
    s2 <> [] /\
    forall m, s1 ++ napp m s2 ++ s3 =~ re.
```

Using induction.

I did not know it was possible to do proof search in the form that it was shown to me.

8:25am. So in summary - fuck `omega`.

I am just so green. I can't believe I could not see all of this despite bashing my head against it for two days. It is not like I've been looking in the wrong place - my headlights were squarely pointed at the culprit and I still could not see it.

I am just so bad at this...but now that I have this thankfully I should be able to make progress.

Let me chill for a bit and then I will start. I will do so by proving the function shown to me on my own."