Yesterday I finished the definitions for filter_challenge and today I will prove it. I'll try the 5 star problem after that.
Right now I am just chilling. I got up early, and I do not feel like doing much else right now.

https://boards.4channel.org/a/thread/187971867/input-incomplete-anime-quotes-let-the-ai-complete

This thread is pretty funny.

Yesterday I tried Dark Tartarus wad, but it is not really to my taste. There is hard, but then there is dropped into a room of 200 monsters with no ammo hard. Or dropped into an open field with 20 archviles hard.

Yeah, I know that there are secrets around the place that would allow the player to work around that and that on easy mode the ammo is plentiful, but still...

I'll give the thing a pass.

It is not like the Coq problems that I get stuck on.

Getting stuck on some math problem is a matter of life and death for me - if I cannot overcome the easy challenges, I will have that much more difficulty with the hard ones.

So I am literally threatened unless I make progress.

I find it really ironic for what reason I am actually studying this. Usually Coq is a tool for PL researchers, but I am picking it up for a diametrically opposite reason - to level up my math skills directly.

I expect math skill to be like poker skill - it is universally recognized that online players are much better than live players.

Math won't be an exception - I will go through SF, the stuff by Adam Chipatala and then I will go through math from the beginning. I am going to need real analysis in order to understand how reals are constructred. All those subjects that I would not give a passing glance earlier are suddenly a priority for me.

But after I do all the fundamentals of both CS and general math, there really won't be anything else after that. I will push forward and start making machine leanring theorems and proving them.

Up to now, I've been working from a different direction that most people would. I spent so much time thinking about omnipotence. Then came the self improvement loop. Then came machine learning and Spiral where I assumed that I had the algorithm and all I needed to do was implement it.

Everything that I've done so far was top down.

Theorem proving that I am doing now is bottom up. I will push forward from here and finally connect with the work that I've did during the past.

I guess after all - it is just too unrealistic to just wait for the research community to give me what I want. If I leave it to them, they really might take decades to get anywhere at this rate.

I might not have any reason to think I will fare better myself in this endeavor, but there is no doubt in my mind on what needs to be done.

There might be a mountain of work. And there is a very real chance that I will miss the insight that I seek.

But because thing are that way - that is precisely the reason why I need to do this. If it is unappealing to me, then it will be unappealing to others. So those with the desire need to step up and accept that responsibility.

This is what my life has been leading up to.

This is what it means to live in reality. I might not have chosen to be born, but regardless of that the universe is looking for those willing to take responsibility for their own existence.