"Let me finally start work for the day. I had a bit of fun slacking off, but now it is time to get a bit done. I've been killing myself trying to prove the pumping lemma for the past few days so a little bit of leisure should be fine, but it needs to be within limits.
It is time to deal with the `filter_challenge`. I think I will manage to find my way through it The 5-star problem after it is what I am worried about, but I need to persevere.

Well whether I manage or not is up to fate. The dice has been cast."